+++ server/actions/fetch-all.js
@@ -1,4 +1,6 @@
 export default function fetchAll(req, res, next) {
-  req.hull.enqueue("fetchAllUsers")
-    .then(next, next);
+  req.hull.enqueue("fetchAllUsers", {
+    updated_after: req.query.updated_after,
+    updated_before: req.query.updated_before
+  }).then(next, next);
 }