+++ server/jobs/fetch-users.js
@@ -3,6 +3,7 @@ import moment from "moment";
 import _ from "lodash";
 
 import saveUsers from "./save-users";
+import fetchAllUsers from "./fetch-all-users";
 
 export default function fetchUsers(ctx, payload = {}) {
   const { intercomAgent } = ctx.shipApp;
@@ -11,10 +12,15 @@ export default function fetchUsers(ctx, payload = {}) {
 
   if (!last_updated_at || !moment(last_updated_at).isValid()) {
     if (ctx.ship.private_settings.last_updated_at
-      && moment(ctx.ship.private_settings.last_updated_at).isValid()) {
+      && moment(ctx.ship.private_settings.last_updated_at).isValid()
+      && moment(ctx.ship.private_settings.last_updated_at).isAfter(moment().subtract(1, "day"))) {
       last_updated_at = ctx.ship.private_settings.last_updated_at;
     } else {
-      last_updated_at = moment().subtract(10, "minutes").format();
+      ctx.metric.event({
+        title: "fetchUsers - wrong last_updated_at value, switching to default - 1 day",
+        text: ctx.ship.private_settings.last_updated_at
+      });
+      last_updated_at = moment().subtract(1, "day").format();
     }
   }
 
@@ -23,7 +29,22 @@ export default function fetchUsers(ctx, payload = {}) {
   return intercomAgent.getRecentUsers(last_updated_at, count, page)
     .then(({ users, hasMore }) => {
       const promises = [];
-      if (hasMore) {
+
+      if ((page + 1) * count >= 10000) {
+        if (process.env.ENABLE_FETCH_ALL_FALLBACK) {
+          const updated_after = last_updated_at;
+          const updated_before = moment(_.get(_.last(users), "updated_at"), "X").format();
+          ctx.metric.event({
+            title: "fetchUsers - going to too high page, switching to fetchAllUsers with filtering",
+            text: JSON.stringify({ updated_after, updated_before })
+          });
+          promises.push(fetchAllUsers(ctx, { updated_before, updated_after }));
+        } else {
+          ctx.metric.event({
+            title: "fetchUsers - going to too high page, stopping the fetch"
+          });
+        }
+      } else if (hasMore) {
         promises.push(fetchUsers(ctx, {
           last_updated_at,
           count,
@@ -37,13 +58,13 @@ export default function fetchUsers(ctx, payload = {}) {
 
       return Promise.all(promises)
         .then(() => {
-          if (page !== 1 || _.isEmpty(users)) {
-            return Promise.resolve();
+          if ((page === 1 || page % 5 === 0) && !_.isEmpty(users)) {
+            const newLastUpdatedAt = moment(_.get(_.first(users), "updated_at"), "X").format();
+            return ctx.helpers.updateSettings({
+              last_updated_at: newLastUpdatedAt
+            });
           }
-          const newLastUpdatedAt = moment(_.get(_.first(users), "updated_at"), "X").format();
-          return ctx.helpers.updateSettings({
-            last_updated_at: newLastUpdatedAt
-          });
+          return Promise.resolve();
         });
     });
 }